python/pyproject.toml
[project]
name = "monocr-onnx"
version = "0.1.0"
description = "Universal MonOCR package with ONNX bindings"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "Janakh", email = "janakh@example.com" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
]
dependencies = [
    "onnxruntime>=1.15.0",
    "Pillow>=9.0.0",
    "numpy>=1.21.0",
    "requests>=2.25.0",
    "tqdm>=4.60.0",
    "click>=8.0.0",
    "pdf2image>=1.16.0",
]

[project.urls]
Repository = "https://github.com/janakh/monocr-onnx"
Issues = "https://github.com/janakh/monocr-onnx/issues"

[project.scripts]
monocr = "monocr_onnx.cli:main"
monocr-download = "monocr_onnx.download:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
exclude = [
  "tests",
  "docs",
  "dist",
  "build",
  "model",  # Exclude model directory
]
